skynet/requirements.txt
# This is an implicit value, here for clarity
--index https://pypi.python.org/simple/

Werkzeug==0.10.4
Jinja2==2.8.0
#
Flask==0.10.1
Flask-Bootstrap==3.3.5.7
Flask-Mail==0.9.1
Flask-Migrate==1.6.0
Flask-Moment==0.5.1
Flask-SQLAlchemy==2.1
Flask-Script==2.0.5
Flask-WTF==0.12.0
Flask-Restful==0.3.4
Flask-Login==0.3.2
#
migrate==0.3.7
SQLAlchemy==1.0.9
WTForms==2.0.2
alembic==0.8.3
#
pyzmq==14.7.0
apscheduler== 3.0.4
elasticsearch==2.1.0
#
pyjwt==1.4.0
cryptography ==1.1
#pycrypto==2.6.1
#ecdsa==0.13